+++ scripts/irccat.py
@@ -12,7 +12,7 @@ import argparse
 import itertools
 
 import irc.client
-import irc.logging
+import jaraco.logging
 
 target = None
 "The nick or channel to which to send messages"
@@ -45,14 +45,14 @@ def get_args():
     parser.add_argument('nickname')
     parser.add_argument('target', help="a nickname or channel")
     parser.add_argument('-p', '--port', default=6667, type=int)
-    irc.logging.add_arguments(parser)
+    jaraco.logging.add_arguments(parser)
     return parser.parse_args()
 
 def main():
     global target
 
     args = get_args()
-    irc.logging.setup(args)
+    jaraco.logging.setup(args)
     target = args.target
 
     reactor = irc.client.Reactor()